chore: add more helpers and improve some test coverage
