+++ packages/vault/src/vault/utils/index.ts
@@ -1,4 +1,12 @@
+export * from './get-balance'
 export * from './get-current-wallet'
+export * from './get-wallet-account-info'
+export * from './get-wallet-transactions'
+export * from './restart-wallet'
+export * from './restore-wallet'
+export * from './sign'
+export * from './submit-tx'
+export * from './switch-network'
 export * from './sync-account-info'
 export * from './sync-transactions'
 export * from './sync-wallet'
