+++ packages/vault/test/network-info/network-info-store.test.ts
@@ -8,14 +8,17 @@ import { DEFAULT_NETWORK_INFO } from '../../src/network-info/default'
 describe('CredentialStore', () => {
   let networkNameMainnet: string
   let networkNameBerkeley: string
+  let networkNameRopsten: string
   let providerConfigMainnet: ProviderConfig
   let providerConfigBerkeley: ProviderConfig
+  let providerConfigRopsten: ProviderConfig
 
   const mockUrl = 'https://...'
   const networkType = 'testnet'
 
   beforeEach(() => {
     networkNameMainnet = 'Mainnet'
+    networkNameRopsten = 'Ropsten'
     // don't use the same network name
     networkNameBerkeley = 'Berkeley Other'
     providerConfigMainnet = {
@@ -43,6 +46,19 @@ describe('CredentialStore', () => {
       networkName: networkNameBerkeley,
       networkType: networkType,
       chainId: '...'
+    }
+    providerConfigRopsten = {
+      nodeEndpoint: {
+        providerName: 'eth-node',
+        url: 'https://ropsten.ethereum.org'
+      },
+      archiveNodeEndpoint: {
+        providerName: 'eth-archive',
+        url: 'https://ropsten.archive.ethereum.org'
+      },
+      networkName: networkNameRopsten,
+      networkType: networkType,
+      chainId: '3'
     }
   })
 
@@ -109,5 +125,52 @@ describe('CredentialStore', () => {
     expect(chainIds.length).toEqual(
       Object.keys(DEFAULT_NETWORK_INFO).length + 2
     )
+  })
+  it('should update existing network info', () => {
+    const { result } = renderHook(() => useVault())
+    act(() => {
+      result.current.setNetworkInfo(networkNameRopsten, providerConfigRopsten)
+      result.current.updateNetworkInfo(networkNameRopsten, { chainId: '4' })
+    })
+    const updatedNetworkInfo = result.current.getNetworkInfo(networkNameRopsten)
+    expect(updatedNetworkInfo.chainId).toEqual('4')
+  })
+
+  it('should not update non-existing network info', () => {
+    const { result } = renderHook(() => useVault())
+    act(() => {
+      result.current.updateNetworkInfo('NonExisting', { chainId: '10' })
+    })
+    const updatedNetworkInfo = result.current.getNetworkInfo('NonExisting')
+    expect(updatedNetworkInfo).toBeUndefined()
+  })
+
+  it('should handle setting and getting the current network info', () => {
+    const { result } = renderHook(() => useVault())
+    act(() => {
+      result.current.setNetworkInfo(networkNameRopsten, providerConfigRopsten)
+      result.current.setCurrentNetworkName(networkNameRopsten)
+    })
+    const currentNetworkInfo = result.current.getCurrentNetworkInfo()
+    expect(currentNetworkInfo).toEqual(providerConfigRopsten)
+  })
+
+  it('should return undefined for non-existent current network', () => {
+    const { result } = renderHook(() => useVault())
+    act(() => {
+      result.current.setCurrentNetworkName('NonExistentNetwork')
+    })
+    const currentNetworkInfo = result.current.getCurrentNetworkInfo()
+    expect(currentNetworkInfo).toBeUndefined()
+  })
+
+  it.skip('should clear all network info', () => {
+    const { result } = renderHook(() => useVault())
+    act(() => {
+      result.current.setNetworkInfo(networkNameRopsten, providerConfigRopsten)
+      result.current.clear()
+    })
+    const networks = result.current.allNetworkInfo()
+    expect(networks.length).toEqual(0)
   })
 })
